+++ src/Controller/CallbackController.php
@@ -91,6 +91,22 @@ class CallbackController extends StorefrontController
     }
 
     #[Route(
+        path: '/altapay/redirect',
+        name: 'altapay.gateway.redirect',
+        defaults: ['auth_required' => false],
+        methods: ['POST']
+    )]
+    public function gatewayRedirect(Request $request, SalesChannelContext $salesChannelContext): Response
+    {
+        $data = [
+            'languageCode' => $request->get('language'),
+            'title' => $this->translator->trans('altapay.gateway.title'),
+        ];
+
+        return $this->render('@WexoAltaPay/gateway/redirect.html.twig', ['gatewayData' => $data]);
+    }
+
+    #[Route(
         path: '/altapay/getError',
         name: 'altapay.gateway.error',
         defaults: ['auth_required' => false],
@@ -116,6 +132,7 @@ class CallbackController extends StorefrontController
 
     #[Route(
         path: '/altapay/callback/notification',
+        name: 'altapay.gateway.notification',
         defaults: ['auth_required' => false],
         methods: ['POST']
     )]
@@ -126,13 +143,12 @@ class CallbackController extends StorefrontController
         }
         try {
             $result = new SimpleXMLElement($request->get('xml'));
-            $orderNumber = (string)$result?->APIResponse?->Body?->Transactions?->Transaction[0]?->ShopOrderId;
-            $paymentId = (string)$result?->APIResponse?->Body?->Transactions?->Transaction[0]?->PaymentId;
+            $orderNumber = (string)$result?->Body?->Transactions?->Transaction?->ShopOrderId;
             if (!$orderNumber) {
                 throw new Exception();
             }
         } catch (Exception) {
-            return new Response('Invalid request', 400);
+            return new Response('Error processing request', 400);
         }
         $criteria = (new Criteria())
             ->addFilter(new EqualsFilter('orderNumber', $orderNumber));
@@ -140,12 +156,12 @@ class CallbackController extends StorefrontController
             ->addSorting(new FieldSorting('createdAt', FieldSorting::DESCENDING));
         /** @var OrderEntity|null $order */
         $order = $this->orderRepository->search($criteria, $salesChannelContext->getContext())->first();
-        if (!$order ||
-            ($order->getCustomFields()[PaymentService::ALTAPAY_PAYMENT_ID_CUSTOM_FIELD] ?? null) !== $paymentId) {
-            return new Response('Invalid request', 400);
+        if (!$order) {
+            return new Response('Order not found in the CMS', 400);
         }
         $transaction = $order->getTransactions()->first();
-        $this->paymentService->transactionCallback($result, $order, $transaction, $salesChannelContext);
-        return new Response(null, 204);
+        $allRequestParams = array_merge($request->query->all(), $request->request->all());
+        $this->paymentService->transactionCallback($result, $order, $transaction, $salesChannelContext, $allRequestParams, true);
+        return new Response("Acknowledged", 200);
     }
 }